Finished the gravity based box ball I built by mistake
